feat: move compile SCV file to server side
