+++ src/services/gererateCSV.ts
@@ -1,7 +1,7 @@
+'use server';
 import { headerCharacters } from '../constants/constants';
 import { type FullCharacterData } from './api';
-
-export function downloadCSV(data: FullCharacterData[], count: number) {
+export async function generateCSV(data: FullCharacterData[]) {
   const rows = data.map((char) => [
     char.name,
     char.eye_color,
@@ -17,13 +17,5 @@ export function downloadCSV(data: FullCharacterData[], count: number) {
       row.map((field) => `"${String(field).replace(/"/g, '""')}"`).join(',')
     )
     .join('\n');
-  const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = `${count}_items.csv`;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(url);
+  return csvContent;
 }
